# packages/api-contracts/openapi/workplace.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/media/uploads/presign": {
      "post": {
        "description": "Returns presigned-style upload metadata.",
        "operationId": "MediaController_createUploadContract",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadUrlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload contract payload containing a scoped upload URL"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create upload contract",
        "tags": [
          "media"
        ]
      }
    },
    "/media/uploads/complete": {
      "post": {
        "operationId": "MediaController_completeUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Validate upload and enqueue malware scan",
        "tags": [
          "media"
        ]
      }
    },
    "/media/{mediaId}/download": {
      "get": {
        "operationId": "MediaController_createDownloadUrl",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a scoped private media download URL",
        "tags": [
          "media"
        ]
      }
    },
    "/media/{mediaId}": {
      "get": {
        "operationId": "MediaController_getMediaStatus",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get private media scan status",
        "tags": [
          "media"
        ]
      }
    },
...
    "/groups": {
      "get": {
        "operationId": "GroupsController_listGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      },
      "post": {
        "operationId": "GroupsController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}/join": {
      "post": {
        "operationId": "GroupsController_joinGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}/requests": {
      "post": {
        "operationId": "GroupsController_requestMembership",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestGroupMembershipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      },
      "get": {
        "operationId": "GroupsController_listMembershipRequests",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}/requests/{requestId}/resolve": {
      "post": {
        "operationId": "GroupsController_resolveMembershipRequest",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveGroupMembershipRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}/invitations": {
      "post": {
        "operationId": "GroupsController_createInvitation",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/groups/invitations/{invitationId}/respond": {
      "post": {
        "operationId": "GroupsController_respondToInvitation",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondGroupInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}/members/{memberUserId}/role": {
      "patch": {
        "operationId": "GroupsController_updateMemberRole",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}/favorite": {
      "patch": {
        "operationId": "GroupsController_setFavorite",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupFavoriteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}/visit": {
      "post": {
        "operationId": "GroupsController_recordVisit",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}": {
      "patch": {
        "operationId": "GroupsController_updateGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      },
      "delete": {
        "operationId": "GroupsController_deleteGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      },
      "get": {
        "operationId": "GroupsController_getGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/groups/onboarding/initialize": {
      "post": {
        "operationId": "GroupsController_initializeStarterGroups",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializeStarterGroupsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{groupId}/members": {
      "get": {
        "operationId": "GroupsController_listMembers",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
...
  },
  "info": {
    "title": "Workplace API",
    "description": "Workplace backend API documentation",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUploadUrlDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "Original client filename"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp",
              "video/mp4",
              "application/pdf",
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
              "application/vnd.openxmlformats-officedocument.presentationml.presentation"
            ]
          },
          "sizeBytes": {
            "type": "number",
            "minimum": 1,
            "maximum": 52428800
          },
          "groupId": {
            "type": "string",
            "description": "Optional group scope for a collaboration upload",
            "format": "uuid"
          }
        },
        "required": [
          "fileName",
          "contentType"
        ]
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "mediaId"
        ]
      },
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {}
      },
      "RequestGroupMembershipDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 280
          }
        }
      },
      "ResolveGroupMembershipRequestDto": {
        "type": "object",
        "properties": {}
      },
      "CreateGroupInvitationDto": {
        "type": "object",
        "properties": {
          "invitedUserId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "invitedUserId"
        ]
      },
      "RespondGroupInvitationDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateGroupMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "member"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "UpdateGroupFavoriteDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {}
      },
      "DeleteGroupDto": {
        "type": "object",
        "properties": {}
      },
      "InitializeStarterGroupsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}